inline websocket connection
