+++ packages/adapters/eth/src/lib.rs
@@ -12,10 +12,10 @@ pub use alloy::primitives::Address;
 pub use aws::*;
 pub use blob_encoder::BlobEncoder;
 pub use error::{Error, Result};
-pub use failover_client::FailoverClient;
+pub use failover_client::{FailoverClient, ProviderConfig};
 pub use http::Provider as HttpClient;
-pub use provider::L1Provider;
 pub use std::sync::Arc;
 pub use websocket::{
     AcceptablePriorityFeePercentages, L1Key, L1Keys, Signer, Signers, TxConfig, WebsocketClient,
+    WebsocketClientFactory,
 };
